Added "More formulas" link to max prepared spells dialog.
Extracted sample formulas to a function in `formula.ts`.
