+++ src/types/types.ts
@@ -231,3 +231,8 @@ export type ExpandedItemIdToLocationsMap = Map<string, Set<string>>;
  * A map from key Item ID to pre-fetched chat data.
  */
 export type ExpandedItemData = Map<string, ItemChatData>;
+
+export type MaxPreparedSpellFormula = {
+  label: string;
+  value: string;
+};
